requirements.txt
python-telegram-bot==20.7
pandas==2.0.3
PyYAML==6.0
Jinja2==3.1.2